+++ CHANGES.md
@@ -1,10 +1,11 @@
 ## Change Log
 
 #### 0.2.4
-* Updates 'rapidminer-release' plugin to version 0.1.6
+* Updates 'rapidminer-release' plugin to version 0.2.0
 * Replaces 'rapidminer-publish' plugin by 'maven-publish' plugin
 * Adds JUnit, nebula-test and Spock as test dependencies
 * Changes release tasks to build and publish
+* Use default Gradle build dir 'build/' instead of 'target/'
 
 #### 0.2.3
 * Updates 'rapidminer-code-quality' plugin to version 0.2.1